docs(docs/*): Docs refactored to enable static site generation.
The documentation is now a combination of Markdown and Jinja2 syntaxes. This
allows devs to have more fine-tuned control over how they are presented.

In this case, Jinja will render the docs into HTML -- injecting any variables
and rendering based on templating logic -- from Markdown which serves as the
primary markup syntax.

Breaking changes: Lots of broken links on SO.