refactor(Gruntfile.coffee): Denoted variable paths
Paths (such as for dist and build) are moved to variables
in the Gruntfile (in the paths namespace), and underscores
are prefixed to their names to make it easier to see them
when developing. .gitignore has been updated to reflect this.

Also, I snuck in the removal of some files not needed for testing
into this commit.