fix(client/js/dnd.js): fixes #953
The reset internal API method for the DnD module was
broken in that it failed with an unchecked error
if no drop zones were configured.  Closer inspection
revealed that an internal variable tracking all drop zones
internally was being overwritten every time a new
drop zone was registered.  This was not the direct
cause of this issue, but fixing this also addressed the
original issue here.