fix(iframe.xss.response.js): Allow JSON with more than one level
The original regex used to parse the response in an iframe in IE9 and younger
was not producing  valid JSON for responses with at least
two levels of nesting.

The matching code has also been refactored out into a separate function,
`readIframeResponse` to help with testability. The code to send the postMessage
to the parent window is still located within a self-executing anonymous func.

Fixes #913