fix(client/js/juqery-plugin.js): #1017 - better support for non-jQuery callback syntax w/ jQuery plug-in
We were already most of the way there, but some logic had
to be adjusted.  Previously, if you included traditional
(non-jQuery) callbacks in your code when using the jQuery
plug-in, returning false from these callback functions would
be ignored.  This would be a problem if you wanted to, for
example, stop a specific file from being submitted by
returning false in your onValidate callback function.  This
should be fixed in this commit.