fix($build): Add jQuery to S3+jQuery builds
The `@fuSrcS3Jquery` modulew was not being prepended with the `@fuSrcJquery` module.
This was a regression caused in 9994e562ef81973bb3c9907cc49a1ab3fa962f09
in order to prevent double loading of modules.

Now, the list of modules are constructed and then filtered to only contain unique
modules (to prevent duplicates).