test($saucelabs): remove iOS7 tests since Sauce Labs sucks ass
iOS7 tests pass locally with the iOS 7 simulator and a real iOS7 device,
but fail in a Sauce Labs VM.  Attempts to run manual tests with Sauce Labs
in an attempt to reproduce are not currently possible for some reason.
Sauce Connect tunnel is up, and Sauce Labs sees the tunnel when starting up
the manual session, but when the browsers load, they claim there is no tunnel.
If you are reading this message, avoid Sauce Labs.  It's a nightmare.