fix($scaling): scaled images sometimes fail to start uploading #1061
Internally, FU waited for MegaPixImage to invoke a callback that FU's image module added to the associated canvas' 2D rendering context.  In some cases, this callback function was apparently not available by the time MegaPixImage finished drawing the scaled image onto the canvas.  This is probably caused by garbage collection of the rendering context object instance that FU inserted the callback on.  The callback was switched to the canvas element itself, which seems to fix the issue.