fix($scaling): rethink smallest scaled file first upload logic #1061
This was broken in a number of ways, and I hope these changes represent a better approach.  I don't like the fact that some concepts of scaling have crept out of the saling module and into other modules (like handler.base.js), but perhaps the concept of a "on-demand-generated blob" and blob proxies is abstract enough to make it appropriate for non-scaling modules to deal with these sorts of details.

TODO: Take another serious look at the changes to FU outside of the scaling module and figure out if any of this can or should be moved into the scaling module.