fix($button): file input isn't tall enough and IE requires dbl-click
In order to overcome the fact that a double-click on the "text" portion of
a file input in IE is required to invoke the file chooser, the font-size
is very large, forcing the button to take up the entire container (button)
in all cases.

Another issue dealt with here: an explicit height specified by 3rd-party
code for input elements may make part of the visible upload button
unclickable if the visible/container button is taller than the specified
input height.  So, we ensure the file input takes up the entire height of
the container.  Width is set to 100% as well for good measure.

fixes #1152
fixes #1277