style(ui-updates): background drop zone text if file DnD is supported
Relies on a new optional template attribute, qq-drop-area-text, on the qq-uploader-selector element.  If this is present, the value of that attribute will be displayed as the background text for the uploader if the current browser supports file DnD.  Displaying the message is done entirely with CSS, so this is dependant on the new css files (not supported in the legacy file).

Once a file is added to the list, the message is removed.  This is done for 2 reasons:
 1. Overlaying the drop files text with file elements doesn't look very nice.
 2. The drop zone text interferes with click events.  This is easily addressed with a pointer-events:none style, but pointer events aren't supported in IE10.

 Once the user selects a file, they probably don't need to be reminded that dropping files is possible anyway.

#843
#1297