feat(templating): Add support for the HTML 5.1 <dialog> element
Fine Uploader UI only, of course.  If the element is supported, and is present in the template, it will be used to display modal messages, prompts, and confirms.  Otherwise, we will fall back to alert dialogs.  The user can still override and of the show* methods/options to provide their own modals as well.

#843