feat(new-ui): make spinner more visible in gallery view
This was done by making the thumbnail a bit opaque when the spinner is visible. The opacity of the thumbnail is changed via CSS when the file's LI contains a "qq-in-progress" class. The templating module adds this class to the file container when the spinner is shown, and removes it when the spinner is hidden.

#1297